+++ projects/admin/src/app/components/modal/ttth/model-ttth-dotthi/model-ttth-dotthi.component.ts
@@ -1,54 +1,48 @@
 import { Component, OnInit, AfterViewInit, OnDestroy } from '@angular/core';
 import { ModalService } from '../../../../services/modal.service';
-import * as ClassicEditor from '@ckeditor/ckeditor5-build-classic';
 import { DotthiService } from '../../../../services/ttth/dotthi.service';
-import { LophocService } from '../../../../services/ttth/lophoc.service';
 import { ttthDotThi } from '../../../../../models/ttthDotThi';
 import { ToastrService } from 'ngx-toastr';
+import { getCookie } from '../../../../../../../common/helper';
+
 @Component({
   selector: 'app-model-ttth-dotthi',
   templateUrl: './model-ttth-dotthi.component.html',
   styleUrls: ['./model-ttth-dotthi.component.css']
 })
 export class ModelTtthDotthiComponent implements OnInit {
-  constructor(private modalService: ModalService,private DotthiService: DotthiService,private toastr: ToastrService,private lophocService: LophocService) { }
+  constructor(private modalService: ModalService,private DotthiService: DotthiService,private toastr: ToastrService) { }
   DotThi: ttthDotThi[];
-  LopHoc: any[];
-
+  private _username: any = getCookie('displayName');
   ngOnInit(): void {
     this.getdanhsach();
-    this.getlophoc();
   }
   closeModal(id: string) {
     this.modalService.close(id)
   }
 
   getdanhsach(): void {
-    this.DotthiService.get().subscribe((data) => {this.DotThi = data; });
-  }
-  getlophoc(): void {
-    this.lophocService.get().subscribe((data) => this.LopHoc = data);
+    this.DotthiService.get().subscribe((data) => {this.DotThi = data;  setTimeout(() => {}, 500);});
   }
   reset():void{
     this.selectedItem=null;
   }
   // add
-  add(tendot: string,lophoc: string,ngaythi: Date): void {
+  add(tendot: string,ngaythi: Date): void {
     const newItem: ttthDotThi = new ttthDotThi();
     newItem.tendot = tendot;
-    newItem.lophoc = lophoc;
     newItem.ngaythi = ngaythi;
     newItem.trangthai = true;
-    newItem.nguoitao = 'hieu';
-    newItem.nguoisua = 'loc';
+    newItem.nguoitao = this._username;
+    newItem.nguoisua = null;
     newItem.created_at = (new Date);
     newItem.updated_at = null;
     this.DotthiService.add(newItem)
       .subscribe(data => {
         this.DotThi.push(data);
-        setTimeout(() => {}, 0);
+        this.getdanhsach();
+
       });
-    this.getdanhsach();
     this.toastr.success('Thêm thành công');
   }
   ///edit
@@ -58,6 +52,7 @@ export class ModelTtthDotthiComponent implements OnInit {
   }
   update(DotThi: ttthDotThi):void {
     DotThi.updated_at= new Date;
+    DotThi.nguoisua= this._username;
     this.DotthiService.update(DotThi)
     .subscribe(data => {
       this.DotThi.push(data);
@@ -66,14 +61,15 @@ export class ModelTtthDotthiComponent implements OnInit {
   }
   //delete
   delete(DotThi: ttthDotThi):void {
-    this.DotthiService.delete(DotThi)
-    .subscribe(data => {
-      this.DotThi.push(data);
-      setTimeout(() => {}, 0);
-    });
-    this.getdanhsach();
-    // window.location.reload();
-    this.toastr.success('Xóa thành công');
+    var comfirmDel = confirm('Bạn có chắc chắn muốn xóa');
+      if(comfirmDel==true){
+      this.DotthiService.delete(DotThi)
+      .subscribe(data => {
+        this.DotThi.push(data);
+        this.getdanhsach();
+      });
+      this.toastr.success('Xóa thành công');
+    }
   }
 
 }
